move _reduced_row_echelon to math
